create_truth_texts/requirements.txt
numpy==1.24.4
opencv-python==4.10.0.82
pillow==10.3.0
pip==20.2.1
setuptools==49.2.1
